move git-related interfaces to pkg/scm/git
