+++ pkg/scm/downloaders/empty/noop.go
@@ -2,6 +2,7 @@ package empty
 
 import (
 	"github.com/openshift/source-to-image/pkg/api"
+	"github.com/openshift/source-to-image/pkg/scm/git"
 	utilglog "github.com/openshift/source-to-image/pkg/util/glog"
 )
 
@@ -13,8 +14,8 @@ type Noop struct {
 }
 
 // Download is a no-op downloader so that Noop satisfies build.Downloader
-func (n *Noop) Download(config *api.Config) (*api.SourceInfo, error) {
+func (n *Noop) Download(config *api.Config) (*git.SourceInfo, error) {
 	glog.V(1).Info("No source location defined (the assemble script is responsible for obtaining the source)")
 
-	return &api.SourceInfo{}, nil
+	return &git.SourceInfo{}, nil
 }
