+++ pkg/scm/downloaders/file/download.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 
 	"github.com/openshift/source-to-image/pkg/api"
+	"github.com/openshift/source-to-image/pkg/scm/git"
 	"github.com/openshift/source-to-image/pkg/util/fs"
 	utilglog "github.com/openshift/source-to-image/pkg/util/glog"
 )
@@ -18,7 +19,7 @@ type File struct {
 }
 
 // Download copies sources from a local directory into the working directory
-func (f *File) Download(config *api.Config) (*api.SourceInfo, error) {
+func (f *File) Download(config *api.Config) (*git.SourceInfo, error) {
 	config.WorkingSourceDir = filepath.Join(config.WorkingDir, api.Source)
 	source := strings.TrimPrefix(config.Source, "file://")
 
@@ -35,7 +36,7 @@ func (f *File) Download(config *api.Config) (*api.SourceInfo, error) {
 		}
 	}
 
-	return &api.SourceInfo{
+	return &git.SourceInfo{
 		Location:   source,
 		ContextDir: config.ContextDir,
 	}, nil
