--- pkg/scm/git/doc.go
@@ -1,4 +1,0 @@
-// Contains the implementation of a Git interface that allows cloning and checking out
-// of git repositories
-
-package git
