+++ pkg/test/download.go
@@ -4,7 +4,7 @@ import (
 	"net/url"
 	"sync"
 
-	"github.com/openshift/source-to-image/pkg/api"
+	"github.com/openshift/source-to-image/pkg/scm/git"
 )
 
 // FakeDownloader provides a fake downloader interface
@@ -16,12 +16,12 @@ type FakeDownloader struct {
 }
 
 // Download downloads a fake file from the URL
-func (f *FakeDownloader) Download(url *url.URL, target string) (*api.SourceInfo, error) {
+func (f *FakeDownloader) Download(url *url.URL, target string) (*git.SourceInfo, error) {
 	f.mutex.Lock()
 	defer f.mutex.Unlock()
 
 	f.URL = append(f.URL, *url)
 	f.Target = append(f.Target, target)
 
-	return &api.SourceInfo{Location: target, CommitID: "1bf4f04"}, f.Err[url.String()]
+	return &git.SourceInfo{Location: target, CommitID: "1bf4f04"}, f.Err[url.String()]
 }
